+++ gemini/textarea.gemini.js
@@ -4,6 +4,7 @@ import Textarea from '../src/textarea';
 const NAME = 'textarea';
 const THEMES = ['alfa-on-white', 'alfa-on-color'];
 const SIZES = process.env.ALL_SIZES ? ['s', 'm', 'l', 'xl'] : ['m'];
+const TEXT = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ';
 
 const PROP_SETS = [
     {},
@@ -17,6 +18,10 @@ const PROP_SETS = [
     {
         placeholder: 'Textarea',
         disabled: true
+    },
+    {
+        label: 'Comment',
+        maxRows: 7
     }
 ];
 
@@ -42,6 +47,22 @@ geminiReact.suite(NAME, () => {
                         suite
                             .render(template)
                             .capture('plain');
+                    } else if (set.label) {
+                        suite
+                            .render(template)
+                            .capture('plain')
+                            .capture('hovered', function (actions) {
+                                actions.mouseMove(this.renderedComponent);
+                            })
+                            .capture('pressed', function (actions) {
+                                actions.mouseDown(this.renderedComponent);
+                            })
+                            .capture('focused', function (actions) {
+                                actions.mouseUp(this.renderedComponent);
+                            })
+                            .capture('with-text', function (actions, find) {
+                                actions.sendKeys(find('.textarea_has-label .textarea__control'), TEXT.repeat(10));
+                            });
                     } else {
                         suite
                             .render(template)