+++ src/calendar-input/calendar-input.jsx
@@ -312,7 +312,7 @@ class CalendarInput extends React.Component {
                         mask='11.11.1111'
                         size={ this.props.size }
                         type='tel'
-                        pattern='[0-9]*'
+                        pattern='[0-9.]*'
                         label={ this.props.label }
                         placeholder={ this.props.placeholder }
                         hint={ this.props.hint }