+++ src/select/select.jsx
@@ -93,6 +93,8 @@ class Select extends React.Component {
             /** Только для type='item': свойства для компонента [MenuItem](#!/MenuItem) */
             props: Type.object
         })),
+        /** Вставляет попап со списком только если элемент активен */
+        renderPopupOnFocus: Type.bool,
         /** Размер компонента */
         size: Type.oneOf(['s', 'm', 'l', 'xl']),
         /** Уникальный идентификатор блока */
@@ -185,7 +187,8 @@ class Select extends React.Component {
         equalPopupWidth: false,
         options: [],
         placeholder: 'Выберите:',
-        mobileMenuMode: 'native'
+        mobileMenuMode: 'native',
+        renderPopupOnFocus: false
     };
 
     static contextTypes = {
@@ -220,6 +223,15 @@ class Select extends React.Component {
      */
     menu;
 
+    /**
+     * @type {Boolean}
+     * При открытом меню, нажатие на Esc устанавливает значение этой переменной в true
+     * Далее фокус переводится на кнопку. Далее вызывается обработчик handleMenuBlur.
+     * В обработчике закрываем попап, если ожидаем закрытия(this.awaitClosing) или фокус за пределами селекта.
+     * Это нужно, т.к. в случае в renderPopupOnFocus={true} меню исчезнет быстрее, чем сработает onMenuBlur
+     */
+    awaitClosing = false;
+
     componentWillMount() {
         this.setState({
             hasGroup: this.props.options.some(option => !!(option.type === 'group' && !!option.content))
@@ -227,17 +239,17 @@ class Select extends React.Component {
     }
 
     componentDidMount() {
-        if (this.popup) {
-            this.setPopupTarget();
-            this.updatePopupStyles();
+        if (this.isAutoSelectRequired()) {
+            this.selectFirstOption();
         }
+
+        this.setPopupTarget();
+        this.updatePopupStyles();
     }
 
     componentWillReceiveProps(nextProps) {
-        if (this.popup) {
-            this.setPopupTarget();
-            this.updatePopupStyles();
-        }
+        this.setPopupTarget();
+        this.updatePopupStyles();
 
         if (this.state.opened && nextProps.disabled) {
             this.toggleOpened();
@@ -402,11 +414,19 @@ class Select extends React.Component {
         let optionsList = this.renderOptionsList(this.props.options);
         let opened = this.getOpened();
         let value = this.getValue();
+        const { popupIsReady } = this.state;
+        const popupIsVisible = this.props.renderPopupOnFocus
+            ? opened && popupIsReady
+            : opened;
+
+        if (!opened && this.props.renderPopupOnFocus) {
+            return null;
+        }
 
         return (
             <Popup
                 key='popup'
-                ref={ (popup) => { this.popup = popup; } }
+                ref={ this.setPopupRef }
                 for={ this.props.name }
                 className={ cn('popup') }
                 directions={ this.props.directions }
@@ -417,14 +437,14 @@ class Select extends React.Component {
                 size={ this.props.size }
                 target={ this.state.isMobile ? 'screen' : 'anchor' }
                 header={ this.state.isMobile && this.renderMobileHeader(cn) }
-                visible={ opened }
+                visible={ popupIsVisible }
                 onClickOutside={ this.handleClickOutside }
                 minWidth={ this.state.popupStyles.minWidth }
                 maxWidth={ this.state.popupStyles.maxWidth }
                 maxHeight={ this.props.maxHeight }
             >
                 <Menu
-                    ref={ (menu) => { this.menu = menu; } }
+                    ref={ this.setMenuRef }
                     className={ cn('menu') }
                     size={ this.props.size }
                     disabled={ this.props.disabled }
@@ -580,7 +600,8 @@ class Select extends React.Component {
     handleMenuBlur(event) {
         event.target.value = this.getValue();
 
-        if (event.relatedTarget !== this.button.getNode()) {
+        if (this.awaitClosing || event.relatedTarget !== this.button.getNode()) {
+            this.awaitClosing = false;
             this.setState({
                 opened: false
             });
@@ -698,9 +719,7 @@ class Select extends React.Component {
                 break;
             case keyboardCode.ESCAPE:
                 event.preventDefault();
-                this.setState({
-                    opened: false
-                });
+                this.awaitClosing = true;
                 this.button.focus();
                 break;
         }
@@ -744,10 +763,8 @@ class Select extends React.Component {
         this.setState({
             isMobile: isMatched
         }, () => {
-            if (this.popup) {
-                this.setPopupTarget();
-                this.updatePopupStyles();
-            }
+            this.setPopupTarget();
+            this.updatePopupStyles();
         });
     }
 
@@ -756,6 +773,34 @@ class Select extends React.Component {
         this.setState({
             opened: false
         });
+    }
+
+    @autobind
+    setPopupRef(ref) {
+        this.popup = ref;
+
+        if (this.popup) {
+            this.popup.setTarget(this.button.getNode());
+        }
+
+        if (this.props.renderPopupOnFocus) {
+            const popupIsReady = !!this.popup;
+
+            this.setState({
+                popupIsReady
+            });
+
+            if (popupIsReady) {
+                setTimeout(() => {
+                    this.focusOnMenu();
+                }, 0);
+            }
+        }
+    }
+
+    @autobind
+    setMenuRef(menu) {
+        this.menu = menu;
     }
 
     /**
@@ -812,6 +857,10 @@ class Select extends React.Component {
     }
 
     focusOnMenu() {
+        if (!this.menu) {
+            return;
+        }
+
         if (this.state.isMobile && this.props.mobileMenuMode === 'popup') return;
 
         let scrollContainer = this.getScrollContainer();
@@ -855,7 +904,7 @@ class Select extends React.Component {
         this.setState({
             opened: newOpenedState
         }, () => {
-            if (newOpenedState && this.menu) {
+            if (newOpenedState) {
                 this.focusOnMenu();
             }
         });
@@ -875,7 +924,9 @@ class Select extends React.Component {
 
     @autobind
     setPopupTarget() {
-        this.popup.setTarget(this.button.getNode());
+        if (this.popup) {
+            this.popup.setTarget(this.button.getNode());
+        }
     }
 
     getCheckedItems(options) {
@@ -917,6 +968,39 @@ class Select extends React.Component {
      */
     getScrollContainer() {
         return this.context.positioningContainerElement || document.body;
+    }
+
+    isAutoSelectRequired() {
+        const { mode, options, renderPopupOnFocus } = this.props;
+
+        return (
+            renderPopupOnFocus &&
+            mode === 'radio' &&
+            options.length > 0 &&
+            !this.hasCheckedItems()
+        );
+    }
+
+    hasCheckedItems() {
+        const { options } = this.props;
+        const checkedItems = this.getCheckedItems(options);
+
+        return checkedItems.length > 0;
+    }
+
+    selectFirstOption() {
+        const firstOption = this.getFirstOption(this.props.options);
+        this.handleOptionCheck([firstOption.value]);
+    }
+
+    getFirstOption(options) {
+        const firstOption = options[0];
+
+        if (firstOption.type === 'group') {
+            return this.getFirstOption(firstOption.content);
+        }
+
+        return firstOption;
     }
 }